pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "latent-intercept"
version = "0.1.0"
description = "Dynamic manipulation via latent world models — Franka arm intercepts a moving block using TD-MPC2 + Genesis"
authors = [{ name = "Shrirag", email = "" }]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "gymnasium>=0.29.0",
    "tensorboard>=2.15.0",
    "matplotlib>=3.8.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "tensordict>=0.3.0",
    "torchrl>=0.3.0",
    "genesis-world",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pytest"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
